requirements.txt
streamlit>=1.36
supabase>=2.4.0,<3
python-dotenv>=1.0.1
pandas>=2.0.0
bcrypt>=4.0.0
